The timings of the performance unit test have dropped for both tests, with and without alterations. Without alterations the improvement is small - in the region of 50ms. With alterations, the improvement is larger, around 200ms. Going from ~750ms to ~700ms and from ~800ms to ~600ms, respectively, at least running them locally.